restructuring of files for lighting tab
