+++ src/components/sections/lighting/subTabs/Sampler.vue
@@ -1,42 +1,23 @@
-<template>
-  <MainTabContent :no-left-pad=false>
-    <ContentBox title="Bank">
-      <ButtonList title="Buttons">
-        <PushButton label="A" :is-active="activeBank === 'SamplerSelectA'" @click="activeBank = 'SamplerSelectA'"/>
-        <PushButton label="B" :is-active="activeBank === 'SamplerSelectB'" @click="activeBank = 'SamplerSelectB'"/>
-        <PushButton label="C" :is-active="activeBank === 'SamplerSelectC'" @click="activeBank = 'SamplerSelectC'"/>
-      </ButtonList>
-      <ColourBox id="colour_one" title="Active / Loaded" :colour-value="getColour('colour_one')"
-                 @colour-changed="onColourChange"/>
-      <ColourBox id="colour_three" title="Sample Empty" :colour-value="getColour('colour_three')"
-                 @colour-changed="onColourChange"/>
-      <ButtonList title="Inactive Bank">
-        <PushButton label="Dim Active Colour" :is-active="isInactiveState('Dimmed')"
-                    @click="setInactiveState('Dimmed')"/>
-        <PushButton label="Inactive Colour" :is-active="isInactiveState('Colour2')"
-                    @click="setInactiveState('Colour2')"/>
-        <PushButton label="Dim Inactive Colour" :is-active="isInactiveState('DimmedColour2')"
-                    @click="setInactiveState('DimmedColour2')"/>
-      </ButtonList>
-      <ColourBox id="colour_two" title="Inactive Bank" :colour-value="getColour('colour_two')"
-                 @colour-changed="onColourChange"/>
-    </ContentBox>
-  </MainTabContent>
-</template>
-
 <script>
 import MainTabContent from "@/components/design/MainTabContent";
 import ContentBox from "@/components/ContentBox";
 import ButtonList from "@/components/button_list/ButtonList";
 import PushButton from "@/components/button_list/Button";
-import {store} from "@/store";
-import {websocket} from "@/util/sockets";
 import ColourBox from "@/components/sections/lighting/ColourBox";
-import {isDeviceMini} from "@/util/util";
+
+import { store } from "@/store";
+import { websocket } from "@/util/sockets";
+import { isDeviceMini } from "@/util/util";
 
 export default {
   name: "LightingSampler",
-  components: {ColourBox, PushButton, ButtonList, ContentBox, MainTabContent},
+  components: {
+    ColourBox,
+    PushButton,
+    ButtonList,
+    ContentBox,
+    MainTabContent,
+  },
 
   data() {
     return {
@@ -84,6 +65,32 @@ export default {
 }
 </script>
 
-<style scoped>
+<template>
+  <MainTabContent :no-left-pad=false>
+    <ContentBox title="Bank">
+      <ButtonList title="Buttons">
+        <PushButton label="A" :is-active="activeBank === 'SamplerSelectA'" @click="activeBank = 'SamplerSelectA'"/>
+        <PushButton label="B" :is-active="activeBank === 'SamplerSelectB'" @click="activeBank = 'SamplerSelectB'"/>
+        <PushButton label="C" :is-active="activeBank === 'SamplerSelectC'" @click="activeBank = 'SamplerSelectC'"/>
+      </ButtonList>
+      <ColourBox id="colour_one" title="Active / Loaded" :colour-value="getColour('colour_one')"
+                 @colour-changed="onColourChange"/>
+      <ColourBox id="colour_three" title="Sample Empty" :colour-value="getColour('colour_three')"
+                 @colour-changed="onColourChange"/>
+      <ButtonList title="Inactive Bank">
+        <PushButton label="Dim Active Colour" :is-active="isInactiveState('Dimmed')"
+                    @click="setInactiveState('Dimmed')"/>
+        <PushButton label="Inactive Colour" :is-active="isInactiveState('Colour2')"
+                    @click="setInactiveState('Colour2')"/>
+        <PushButton label="Dim Inactive Colour" :is-active="isInactiveState('DimmedColour2')"
+                    @click="setInactiveState('DimmedColour2')"/>
+      </ButtonList>
+      <ColourBox id="colour_two" title="Inactive Bank" :colour-value="getColour('colour_two')"
+                 @colour-changed="onColourChange"/>
+    </ContentBox>
+  </MainTabContent>
+</template>
+
+<style>
 
 </style>
